LIBFCREPO-739. Added completed item logging for delete.
https://issues.umd.edu/browse/LIBFCREPO-739

+++ plastron/commands/delete.py
@@ -1,5 +1,8 @@
 import logging
-from plastron.util import get_title_string, ResourceList, parse_predicate_list
+from email.utils import parsedate_to_datetime
+
+from plastron.exceptions import FailureException, RESTAPIException
+from plastron.util import get_title_string, ResourceList, parse_predicate_list, ItemLog
 
 logger = logging.getLogger(__name__)
 
@@ -27,6 +30,11 @@ def configure_cli(subparsers):
         dest='use_transactions'
     )
     parser.add_argument(
+        '--completed',
+        help='file recording the URIs of deleted resources',
+        action='store'
+    )
+    parser.add_argument(
         '-f', '--file',
         help='File containing a list of URIs to delete',
         action='store'
@@ -47,6 +55,19 @@ class Command:
         if self.dry_run:
             logger.info('Dry run enabled, no actual deletions will take place')
 
+        if args.completed:
+            logger.info(f'Reading the completed items log from {args.completed}')
+            # read the log of completed items
+            fieldnames = ['uri', 'title', 'timestamp']
+            try:
+                self.completed = ItemLog(args.completed, fieldnames, 'uri')
+                logger.info(f'Found {len(self.completed)} completed item(s)')
+            except Exception as e:
+                logger.error(f"Non-standard map file specified: {e}")
+                raise FailureException()
+        else:
+            self.completed = []
+
         resources = ResourceList(
             repository=self.repository,
             uri_list=args.uris,
@@ -59,9 +80,20 @@ class Command:
         )
 
     def delete_item(self, resource, graph):
+        if resource.uri in self.completed:
+            logger.info(f'Resource {resource.uri} has already been deleted; skipping')
+            return
         title = get_title_string(graph)
         if self.dry_run:
             logger.info(f'Would delete resource {resource} {title}')
         else:
-            self.repository.delete(resource.uri)
-            logger.info(f'Deleted resource {resource} {title}')
+            response = self.repository.delete(resource.uri)
+            if response.status_code == 204:
+                logger.info(f'Deleted resource {resource} {title}')
+                self.completed.append({
+                    'uri': resource.uri,
+                    'title': title,
+                    'timestamp': parsedate_to_datetime(response.headers['date']).isoformat('T')
+                })
+            else:
+                raise RESTAPIException(response)
